split up lib into modules
